chore: add additional template parameters
