+++ src/Uno.Extensions.Templates/content/unoapp-extensions/Directory.Packages.props
@@ -17,18 +17,26 @@
     <PackageVersion Include="Uno.Core.Extensions.Logging.Singleton" Version="4.0.1" />
     <PackageVersion Include="Uno.Extensions.Authentication" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Authentication.WinUI" Version="255.255.255.255" />
+    <!--#if (configuration)-->
     <PackageVersion Include="Uno.Extensions.Configuration" Version="255.255.255.255" />
+    <!--#endif-->
     <PackageVersion Include="Uno.Extensions.Core" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Hosting" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Hosting.WinUI" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Http" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Http.Refit" Version="255.255.255.255" />
+    <!--#if (localization)-->
     <PackageVersion Include="Uno.Extensions.Localization" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Localization.WinUI" Version="255.255.255.255" />
+    <!--#endif-->
+    <!--#if (uselogging)-->
     <PackageVersion Include="Uno.Extensions.Logging.OSLog" Version="1.4.0" />
+    <!--#if (useserilog)-->
     <PackageVersion Include="Uno.Extensions.Logging.Serilog" Version="255.255.255.255" />
+    <!--#endif-->
     <PackageVersion Include="Uno.Extensions.Logging.WebAssembly.Console" Version="1.4.0" />
     <PackageVersion Include="Uno.Extensions.Logging.WinUI" Version="255.255.255.255" />
+    <!--#endif-->
     <PackageVersion Include="Uno.Extensions.Navigation" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Navigation.Toolkit.WinUI" Version="255.255.255.255" />
     <PackageVersion Include="Uno.Extensions.Navigation.WinUI" Version="255.255.255.255" />
@@ -46,6 +54,9 @@
     <PackageVersion Include="Uno.Wasm.Bootstrap" Version="3.3.1" />
     <PackageVersion Include="Uno.Wasm.Bootstrap.DevServer" Version="3.3.1" />
     <PackageVersion Include="Uno.WinUI" Version="4.6.19" />
+    <!--#if (markup)-->
+    <PackageVersion Include="Uno.WinUI.Markup" Version="4.6.0-dev.13" />
+    <!--#endif-->
     <PackageVersion Include="Uno.WinUI.RemoteControl" Version="4.6.19" />
     <PackageVersion Include="Uno.WinUI.Skia.Gtk" Version="4.6.19" />
     <PackageVersion Include="Uno.WinUI.Skia.Linux.FrameBuffer" Version="4.6.19" />
