+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Mobile/Android/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="MyExtensionsApp" android:versionCode="1" android:versionName="1.0">
-  <uses-sdk android:minSdkVersion="19" android:targetSdkVersion="31" />
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.companyname.myextensionsapp" android:versionCode="1" android:versionName="1.0">
+  <uses-sdk android:minSdkVersion="21" android:targetSdkVersion="31" />
   <application android:label="MyExtensionsApp"></application>
 </manifest>
