+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Mobile/MacCatalyst/Info.plist
@@ -7,7 +7,7 @@
 	<key>CFBundleDisplayName</key>
 	<string>MyExtensionsApp</string>
 	<key>CFBundleIdentifier</key>
-	<string>com.companyname.changeme</string>
+	<string>com.companyname.myextensionsapp</string>
 	<key>CFBundleShortVersionString</key>
 	<string>1.0</string>
 	<key>CFBundleVersion</key>
