+++ src/Uno.Extensions.Templates/content/unoapp-extensions/MyExtensionsApp.Mobile/iOS/Info.plist
@@ -5,7 +5,7 @@
 	<key>CFBundleDisplayName</key>
 	<string>MyExtensionsApp</string>
 	<key>CFBundleIdentifier</key>
-	<string>com.companyname.MyExtensionsApp</string>
+	<string>com.companyname.myextensionsapp</string>
 	<key>CFBundleShortVersionString</key>
 	<string>1.0</string>
 	<key>CFBundleVersion</key>
